Move LazilyInitialized and MutexLocked to `@foxglove/den`
