+++ packages/den/async/LazilyInitialized.test.ts
@@ -2,13 +2,13 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import LazilyInitialized from "@foxglove/studio-base/services/CachedRemoteLayoutStorage/LazilyInitialized";
-import signal from "@foxglove/studio-base/util/signal";
+import LazilyInitialized from "./LazilyInitialized";
 
 describe("LazilyInitialized", () => {
   it("calls the provided compute function only once", async () => {
-    const sig = signal<number>();
-    const compute = jest.fn().mockReturnValue(sig);
+    let resolve: undefined | ((_: number) => void);
+    const promise = new Promise<number>((res) => (resolve = res));
+    const compute = jest.fn().mockReturnValue(promise);
     const lazy = new LazilyInitialized(compute);
     expect(compute).toHaveBeenCalledTimes(0);
 
@@ -19,7 +19,7 @@ describe("LazilyInitialized", () => {
     await Promise.resolve();
     expect(compute).toHaveBeenCalledTimes(1);
 
-    sig.resolve(1);
+    resolve!(1);
     expect(await promise1).toBe(1);
     expect(await promise2).toBe(1);
     expect(compute).toHaveBeenCalledTimes(1);
