+++ packages/den/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "@foxglove/den",
+  "description": "Incubation area for code that may be released under future @foxglove packages",
+  "version": "0.0.0",
+  "license": "MPL-2.0",
+  "private": true,
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/foxglove/studio.git"
+  },
+  "author": {
+    "name": "Foxglove Technologies",
+    "email": "support@foxglove.dev"
+  },
+  "homepage": "https://foxglove.dev/",
+  "main": "./src/index.ts",
+  "files": [
+    "dist",
+    "src"
+  ],
+  "scripts": {
+    "prepack": "tsc -b"
+  },
+  "devDependencies": {
+    "typescript": "4.3.2"
+  },
+  "dependencies": {
+    "async-mutex": "0.3.1"
+  }
+}
